+++ apps/web/src/components/SearchRequestTable/SearchRequestsTableApi.stories.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { Meta, StoryFn } from "@storybook/react";
 import { fakeSearchRequests } from "@gc-digital-talent/fake-data";
 
-import SearchRequestsTable from "./SearchRequestsTable";
+import SearchRequestsTable from "./SearchRequestsTableApi";
 
 const requestsData = fakeSearchRequests();
 const mockPaginatorInfo = {
